LinearSet#kleeneStar(): Remove
This function is unused and wrong. It is wrong since, for example, the
Kleene star of ab* would have been a*b*, which is wrong (using regular
expressions here to express linear sets). In fact, the Kleene star of a
linear set is in general just semi-linear, and so this function cannot
be implemented properly.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>