Fix SemilinearSet.kleeneStar() again
The last fix resulted in the Kleene start of [(a), (b)] being [(),
(a)+{(a)}*, (b)+{(b)}*, (ab)+{(ab)}*]. The last entry here is wrong,
because, for example, abb cannot be generated.

This commit fixes this. This replaces (ab)+{(ab)}* with (ab)+{(a),(b)}*,
which is what was actually meant to be computed.

Sadly, the incorrect behaviour was also captured in a unit test, which
is also fixed by this commit.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>